pagination et autre

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -4,19 +4,24 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
    IRestaurantsRepository restaurantsRepository,
-   IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+   IDishesRepository dishesRepository,
+   IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogWarning("Removing all dishes from restaurant: {RestaurantId}", request.RestaurantId);
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+
+        if ((!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))) throw new ForbidException();
 
         await dishesRepository.Delete(restaurant.Dishes);
     }
